Fix crashes related to mod loading order
On Fabric: the plugins were initialized right after the mod. This caused some plugins to be executed before their owning mod were initialized, leading to crashes.

On Forge: the mod was loaded on mod completion instead as a workaround.

The plugins are now registered as late as possible either on opening to config screen, or when joining a world.