Fix color customization affecting banners
The color key system implementation used the same array instances to store colors as the dyes from vanilla Minecraft.

Modifying the value of a color key via config caused the corresponding day color to be modified, thus affecting the color of banner items and blocks.

The color components arrays from dyes are now copied when creating the base color keys.

Closes #105